Move-only: Move ThreadCheckObfuScationPool to masternodeman
