+++ src/masternodeman.h
@@ -157,4 +157,6 @@ public:
     void UpdateMasternodeList(CMasternodeBroadcast mnb);
 };
 
+void ThreadCheckObfuScationPool();
+
 #endif
