+++ src/obfuscation.cpp
@@ -668,40 +668,3 @@ void CObfuscationPool::RelayCompletedTransaction(const int sessionID, const bool
     for (CNode* pnode : vNodes)
         pnode->PushMessage(NetMsgType::DSC, sessionID, error, errorID);
 }
-
-//TODO: Rename/move to core
-void ThreadCheckObfuScationPool()
-{
-    if (fLiteMode) return; //disable all Obfuscation/Masternode related functionality
-
-    // Make this thread recognisable as the wallet flushing thread
-    util::ThreadRename("prcycoin-obfuscation");
-    LogPrintf("Masternodes thread started\n");
-
-    unsigned int c = 0;
-
-    while (true) {
-        MilliSleep(1000);
-
-        // try to sync from all available nodes, one step at a time
-        masternodeSync.Process();
-
-        if (masternodeSync.IsBlockchainSynced()) {
-            c++;
-
-            // check if we should activate or ping every few minutes,
-            // start right after sync is considered to be done
-            if (c % MASTERNODE_PING_SECONDS == 1) activeMasternode.ManageStatus();
-
-            if (c % 60 == 0) {
-                mnodeman.CheckAndRemove();
-                mnodeman.ProcessMasternodeConnections();
-                masternodePayments.CleanPaymentList();
-                CleanTransactionLocksList();
-            }
-
-            obfuScationPool.CheckTimeout();
-            obfuScationPool.CheckForCompleteQueue();
-        }
-    }
-}
